refactor!(tpm2_protocol): drop tpm_response! and TpmParameters
Used the pattern matching mechanism of `macro_rules!`, turning pre-existing
parameters as self-documenting. In addition, a named parameter called
`kind` was added for the sake of  differentiating commands and responses.

The refined macro was also extended to be able to populate the parameter
area.

This was applied to the pre-existing commands and responses as follows:

1. Substituted each `tpm_struct!` with updated declaration.
2. Substituted each `tpm_response!` with a new `tpm_struct!` declaration.
   This removes ambiguity between responses, some of which are using
   `tpm_struct!`, and some using `tpm_response!` depending on whether they
   have a parameter area, or not.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>