fix(data): adjust TPM2B_PRIVATE_KEY_RSA size
Section "11.2.4.7 TPM2B_PRIVATE_KEY_RSA" has two notes.

Note 1:

"All primes are required to have exactly half the number of significant bits
as the public modulus, and the square of each prime is required to have the
same number of significant bits as the public modulus.  End of Note"

Note 2:

"RSA_PRIVATE_SIZE is a vendor specific value that can be (MAX_RSA_KEY_BYTES
/ 2) or ((MAX_RSA_KEY_BYTES * 5) / 2. The larger size would only apply to
keys that have fixedTPM parents.  The larger size was added in version
1.59."

For the time being, act aligning with Note 1, while at the same it is
better to document both notes to this commit message in order to be
transparent if new problems arise in the future.

Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>