fix(message): return `BuildCapacity` from build traits
Build traits incorrectly return `ParseCapacity` when they should return
`BuildCapacity`. Fixup the error code.

Capacity means the value of an architectural limit in this context. E.g.,
the builder and parser detect if a `TPML_*` has more values than the TCG
specifications allows, and return capacity error if the limit is surpassed.

Additionally, updated documentation of these error values to better
clarify their use and their purpose.

`BuildOverflow` is most likely the most confusing of the pre-existing error
codes. It pre-dates to the time when the implementation had internal
buffers. As of today, the only function returning that error is
`TpmWriter::write_bytes`, and when this happens, it is a clear sign of a
regression in the crate implementation i.e., not part of the legit behavior
anymore.

In the main branch `BufferOverflow` will be eventually removed but it is
out of scope for the bug fix.

Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>