feat!(context): ContextKey
Declare `ContextKey` and change cache format also to store public area of
the key. Turn `key` into local command with the help of this type, as the
subcommand can read public area from the offline storage.

Note that this does not affect context refresh, as refresh is done as part
of the initialization, not part of the subcommand.

Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>