refactor(command): re-implement print_table
Tabular output in `tpm2sh` is simple and minimal, and uses only bolding to
enhance the output. The complexity of abstraction in `tabled` does not
really pay off in this application. Further, removing `tabled` reduces the
number of dependencies from 198 to 177.

In order to decrease complexity and build times, resolve the situation
as follows:

1. Remove `tabled` from dependencies.
2. Re-implement `print_table` and use `clap::builder::styling::Style`
   for outputting ANSI codes for bolding.

Signed-off-by: Jarkko Sakkinen <jarkko@kernel.org>